backend/requirements.txt
fastapi==0.109.0
gunicorn==21.2.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
redis==5.0.1
sentry-sdk[fastapi]==1.40.0
cryptography==42.0.0
PyJWT==2.8.0
python-multipart==0.0.6
slowapi==0.1.9
tweepy==4.14.0
linkedin-api==2.2.0
instagrapi==2.0.0
reportlab==4.0.9
pillow==10.2.0
